Rename: change components and pages directory structure for auth route
