+++ src/components/pages/Todo/TodoList.jsx
@@ -1,5 +1,5 @@
 import React, { useContext } from 'react';
-import { TodosContext } from '../../store/todo-context';
+import { TodosContext } from '../../../store/todo-context';
 import TodoItem from './TodoItem';
 import classes from './TodoList.module.css';
