+++ src/components/pages/Todo/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import TodoForm from './TodoForm';
 import TodoList from './TodoList';
-import classes from './Todos.module.css';
+import classes from './index.module.css';
 
 const Todos = () => (
   <div className={classes.todos}>
